virus_scan_project/api-gateway/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.5.2
marshmallow==3.19.0
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
werkzeug==2.2.3
gunicorn==20.1.0
pyyaml==6.0
pytest==7.3.1
pytest-flask==1.2.0
python-magic==0.4.27